# perf(http): avoid checking promise every request