--- a/ext/http/01_http.js
+++ b/ext/http/01_http.js
@@ -181,13 +181,15 @@
     remoteAddr,
     localAddr,
   ) {
+    const isPromise = true;
     return async function respondWith(resp) {
       try {
-        if (ObjectPrototypeIsPrototypeOf(PromisePrototype, resp)) {
+        if (isPromise || ObjectPrototypeIsPrototypeOf(PromisePrototype, resp)) {
           resp = await resp;
         }
 
         if (!(ObjectPrototypeIsPrototypeOf(ResponsePrototype, resp))) {
+          if (!isPromise) resp = await resp;
           throw new TypeError(
             "First argument to respondWith must be a Response or a promise resolving to a Response.",
           );