--- a/ext/http/01_http.js
+++ b/ext/http/01_http.js
@@ -39,7 +39,6 @@
     ArrayPrototypeSome,
     Error,
     ObjectPrototypeIsPrototypeOf,
-    PromisePrototype,
     Set,
     SetPrototypeAdd,
     SetPrototypeDelete,